service-desk-crew/pyproject.toml
[project]
name = "service_desk_crew"
version = "0.1.0"
description = "service_desk_crew using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]==1.11.0",
    "requests>=2.28.0",
    "pyyaml>=6.0",
    "mcp>=1.8.0",
    "httpx>=0.28.1",
]

[tool.hatch.build.targets.wheel]
packages = ["src/service_desk_crew"]

[project.scripts]
service_desk_crew = "service_desk_crew.main:run"
run_crew = "service_desk_crew.main:run"
train = "service_desk_crew.main:train"
replay = "service_desk_crew.main:replay"
test = "service_desk_crew.main:test"
run_with_trigger = "service_desk_crew.main:run_with_trigger"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
